requirements.txt
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
PyYAML>=6.0
playwright>=1.40.0
python-dotenv>=1.0.0
